requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
transformers>=4.15.0
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
web3>=5.24.0
python-dotenv>=0.19.0
tqdm>=4.62.3
matplotlib>=3.4.3
networkx>=2.6.3 